requirements.txt
psycopg2
py-bcrypt
pandas
PyQt5